Add ability to see solution on game over
Suggestion to fix https://github.com/Assios/chessguessr/issues/23.

This adds the solution as a row on top of the guess-grid if you have finished
the game. This allows you to look at the correct solution and the board at the
same time after having closed the game-over-modal.

I had to tweak the margins a bit to get it to look natural, but it seems to be
ok on both desktop and mobile