Allow user to "go forwards" through move history
Suggested solution to fix https://github.com/Assios/chessguessr/issues/35
(and https://github.com/Assios/chessguessr/issues/34).

The meat of the solution is to keep track of what you have guessed before in a
"cache" for both positions and moves. To do this, we wrap the simple
`setCurrentMove` and `setFenHistory` functions in a layer that writes the cache
every time you make a new/different move, but maintains the cache as long as you
play "the same moves as before" – regardless of _how_ you play them.

This allows "redo" if you have accidentally clicked "undo" too many times.
As an added bonus, after you have placed a guess, you can press "right right
right" to fill in the first three moves of your previous guess. Very convenient
if you already have most of the solution and only need to figure out the final
1-2 moves.

For now (to avoid unsolicited clutter), I have not added a visual UI for the
feature, only added the hotkey arrow-right.